remove check_includes = false from brave/browser/profiles
