+++ browser/profiles/sources.gni
@@ -1,0 +1,10 @@
+brave_browser_profiles_sources = [
+    "//brave/browser/profiles/brave_profile_avatar_downloader.cc",
+    "//brave/browser/profiles/brave_profile_avatar_downloader.h",
+    "//brave/browser/profiles/brave_profile_manager.cc",
+    "//brave/browser/profiles/brave_profile_manager.h",
+    "//brave/browser/profiles/brave_renderer_updater.cc",
+    "//brave/browser/profiles/brave_renderer_updater.h",
+    "//brave/browser/profiles/brave_renderer_updater_factory.cc",
+    "//brave/browser/profiles/brave_renderer_updater_factory.h",
+  ]
