+++ browser/search/BUILD.gn
@@ -11,7 +11,7 @@ source_set("search") {
 
     deps += [
       "//base",
-      "//brave/browser/profiles:profiles",
+      "//brave/browser/profiles:util",
       "//brave/components/constants",
       "//chrome/browser/ui",
       "//chrome/common",
