+++ components/brave_shields/content/test/BUILD.gn
@@ -50,6 +50,8 @@ source_set("unit_tests") {
       "brave_shields_util_unittest.cc",
     ]
 
-    deps += [ "//brave/browser/profiles" ]
+    deps += [
+      "//chrome/browser",
+    ]
   }
 }
