+++ renderer/test/BUILD.gn
@@ -23,7 +23,6 @@ source_set("browser_tests") {
   deps = [
     "//base/test:test_support",
     "//brave/browser/brave_wallet",
-    "//brave/browser/profiles:profiles",
     "//brave/common",
     "//brave/components/brave_wallet/browser",
     "//brave/components/brave_wallet/browser:utils",
